requirements.txt
aiohttp==3.8.4
fal_client==0.4.1
fastapi==0.111.0
openai==1.35.7
pydantic==2.7.4
python-dotenv==1.0.1
python-telegram-bot==21.3
SQLAlchemy==2.0.13
typing_extensions==4.12.2
uvicorn==0.30.1
